When adding a favorite url to our db, turn the url into a permalink.
Add in functionality to create a discord permalink from a normal
image url. We also implement worker_threads, and spawn a thread
when we download something. This is to not block the mainthread,
and this is so we can Hash files off of the main thread.

+++ src/utils/fileLoader.ts
@@ -1,38 +1,27 @@
-import { statSync, createWriteStream } from 'fs';
-import got from 'got';
-import { TextChannel } from 'discord.js';
+import { statSync } from 'fs';
+import { TextChannel, Message } from 'discord.js';
 import { bot } from './bot';
+import { downloadFile } from './downloadFile';
 
 /**
- * In a wonderful effort to defeat heroku's ephemeral filesystem and to never
- * touch a real database, this file has 2 functions, export and import
- *
- * Export will upload the file to a special discord channel and delete
- * the previous version of that file
+ * Export will upload a given file to either dbChannel, or a given discord channel,
+ * and delete the previous version of that file on discord (by default, change with deleteOld)
+ * You can also choose to archive said file, which means we upload it to a specific archiveChannel
+ * AND we upload it to dbChannel, we never delete the old version in the archiveChannel.
  *
  * Import will take a fileName and search a discord channel for an attachment
  * that has the same name, and download it to the root directory
- *
  */
 
-async function downloadFile(url: string, fileName: string): Promise<void> {
-   await new Promise((resolve) => {
-      got.stream(url, {
-         headers: {
-            'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.86 Safari/537.36',
-         },
-      })
-         .pipe(createWriteStream(fileName))
-         .on('finish', () => resolve());
-   });
+function getPath(path: string): RegExpMatchArray | void {
+   return (path.match(/.*\/([^\\]+)\//) || undefined);
 }
 
 export async function importFile(fileName: string): Promise<boolean> {
-   const path = fileName.match(/.*\/([^\\]+)\//);
-   if (path) {
+   if (getPath(fileName)) {
       throw new Error('Imported files should not contain filepaths.');
    }
-   const channel = await bot.channels.get(process.env.dbChannel as string) as TextChannel;
+   const channel = bot.channels.get(process.env.dbChannel as string) as TextChannel;
    if (!channel) {
       throw new Error(`Unable to read channel: ${process.env.dbChannel}, Bad Permissions or incorrect env vars`);
    }
@@ -46,36 +35,46 @@ export async function importFile(fileName: string): Promise<boolean> {
    return (file.attachments.first().filesize === statSync(fileName).size);
 }
 
-export async function exportFile(fName: string, isArchive?: boolean): Promise<boolean> {
-   let fileName = fName;
+export async function exportFile(
+   fName: string,
+   doArchive?: boolean,
+   channelId = process.env.dbChannel,
+   deleteOld = true,
+): Promise<string | undefined> {
    // if the fileName contains a path, we remove the path and just get the fileName
-   const path = fileName.match(/.*\/([^\\]+)\//);
-   if (path) {
-      fileName = fileName.slice(path[0].length);
+   if (!channelId) {
+      throw new Error('Error! channelId set incorrectly!');
    }
-   // get the channel and make sure its not undefined
-   const channel = await bot.channels.get(process.env.dbChannel as string) as TextChannel;
+   const path = getPath(fName);
+   const fileName = (path ? fName.slice(path[0].length) : fName);
+   const channel = bot.channels.get(channelId) as TextChannel;
    if (!channel) {
-      throw new Error(`Unable to read channel: ${process.env.dbChannel}, Bad Permissions or incorrect env vars`);
-   }
-   // if we are told to archive the file, we upload it without deleting the old version
-   if (isArchive) {
-      const archive = bot.channels.get(process.env.archiveChannel as string) as TextChannel;
-      if (archive) {
-         await archive.send({ files: [fileName] });
-      }
-   }
-   const msgs = await channel.fetchMessages();
-   const file = msgs.find(m => !!(m.attachments.size
-      && m.attachments.first().filename === fileName));
-   if (!file) {
-      throw new Error(`file: ${fileName} not found`);
+      throw new Error(`Unable to read channel: ${channelId}`);
    }
 
-   // send the new file, so if something crashes the old file is still in the chan
-   await channel.send({ files: [fileName] });
-   // delete the old file
-   const oldMsg = await channel.fetchMessage(file.id);
-   await oldMsg.delete();
-   return true;
+   // if we are told to archive the file, we upload it without deleting the old version
+   if (doArchive) {
+      const archive = bot.channels.get(process.env.archiveChannel as string) as TextChannel;
+      if (archive) {
+         await archive.send({ files: [fName] });
+      }
+   }
+
+   // upload the file first, just in case
+   const uploadedUrl = (await channel.send({ files: [fName] }) as Message).attachments.first().url;
+
+   // if told to delete the old version, search for the fileaName, then delete it
+   if (deleteOld) {
+      const msgs = await channel.fetchMessages();
+      const file = msgs.find(m => !!(m.attachments.size
+         && m.attachments.first().filename === fileName));
+      if (!file) {
+         throw new Error(`file: ${fileName} not found`);
+      }
+      // if found, delete the old file
+      const oldMsg = await channel.fetchMessage(file.id);
+      await oldMsg.delete();
+   }
+
+   return uploadedUrl;
 }
